requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2
python-decouple==3.1
djangorestframework==3.9.1
markdown==3.0.1
django-filter==2.1.0
Pillow==5.4.1
django-imagekit==4.0.2
django-cors-headers==2.5.0
django-hosts==3.0
django-webpack-loader==0.6.0
django-pagseguro2
numpy==1.17.0
celery==4.3.0
django-cleanup==4.0.0
